% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_disconnect}
\alias{bloomberg_disconnect}
\title{bloomberg_disconnect}
\usage{
bloomberg_disconnect()
}
\description{
Close a Bloomberg connection and reset the cache so that future
connection attempts are possible.
}
