% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_query_curve}
\alias{bloomberg_query_curve}
\title{Get a curve from Bloomberg}
\usage{
bloomberg_query_curve(curve_id, max_maturity = NULL, interp = FALSE)
}
\arguments{
\item{curve_id}{The Bloomberg code for the curve of interest}

\item{max_maturity}{Maximum maturity for the underlying instruments}

\item{interp}{A logical. If true linearly interpolated values will be returned for each day}
}
\value{
A data frame of CurveID, Date, Value
}
\description{
Get a curve from Bloomberg
}
\examples{
library(FIRVr)
bloomberg_query_curve("YCSW0133 Index", interp = TRUE, max_maturity = Sys.Date() + 365)
}
