% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_query_dataset}
\alias{bloomberg_query_dataset}
\title{Get a table of data}
\usage{
bloomberg_query_dataset(security, field, options = NULL, overrides = NULL)
}
\arguments{
\item{security}{A single security identifier}

\item{field}{A single field}

\item{options}{An optional named character vector with option values}

\item{overrides}{An optional named character vector with override values}
}
\value{
Data frame of results
}
\description{
Get a table of data
}
\examples{
bloomberg_query_table("RXZ9 Comdty", "FUT_DLVRBLE_BNDS_CUSIPS")
}
