% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_query_intraday}
\alias{bloomberg_query_intraday}
\title{Get a data frame of intraday prices for a list of securities}
\usage{
bloomberg_query_intraday(
  securities,
  field = "open",
  from_datetime = Sys.time() - 60 * 60 * 24 * 60,
  to_datetime = NULL,
  interval = 5,
  timezone = "Europe/London",
  options = NULL
)
}
\arguments{
\item{securities}{A list of Bloomberg identifiers}

\item{field}{The field for which intraday data is to be retrieved}

\item{from_datetime}{Start datetime}

\item{to_datetime}{End datetime}

\item{interval}{Interval of observations}

\item{options}{A named list of options to pass to Bloomberg}
}
\value{
A data frame of tidy data
}
\description{
Get a data frame of intraday prices for a list of securities
}
\examples{
bloomberg_query_intraday("DE0001135176 Govt", field = "open",
  from_datetime = Sys.time() - 60 * 60 * 24 * 60)
}
