% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_query_search}
\alias{bloomberg_query_search}
\title{Get a list of bond IDs from a pre-defined Bloomberg search}
\usage{
bloomberg_query_search(search_name)
}
\arguments{
\item{search_name}{The name of a predefined Bloomberg search}
}
\value{
Data frame of Bloomberg bsrch results
}
\description{
Get a list of bond IDs from a pre-defined Bloomberg search
}
\examples{
bloomberg_query_search("COMDTY:Vessel")
}
