% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bloomberg_query_static}
\alias{bloomberg_query_static}
\title{Query Bloomberg for static data.}
\usage{
bloomberg_query_static(
  securities,
  fields,
  options = NULL,
  overrides = NULL,
  tidy_data = TRUE
)
}
\arguments{
\item{securities}{A list of Bloomberg security IDs}

\item{fields}{Timeseries fields}

\item{options}{An optional named list of options to pass to Bloomberg}

\item{overrides}{An optional names list of overrides to pass to Bloomberg}

\item{tidy_data}{A boolean to specify whether we would like tidy or wide data}
}
\value{
A data frame of tidy data with columns Security, Field as many as required
        of ValueChr, ValueDate, ValueDouble, ValueInt and ValueLgl (logical/boolean).
        If tidy_data is set to FALSE, a wide data frame with one column per field.
}
\description{
Query Bloomberg for static data.
}
\examples{
bloomberg_query_static("DE0001135176 Govt", "ID_ISIN")
}
