% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_basis.R
\name{bond_basis_model}
\alias{bond_basis_model}
\title{Generate a bond basis model. Simulate many possible outcomes for given
input parameters and calculate some derived fields based on each state.}
\usage{
bond_basis_model(
  means,
  stdevs,
  correlations,
  field_descs,
  static_data,
  forward_prices,
  future_price,
  evaluation_date = NULL,
  n_periods = 63,
  n_simulations = 1000
)
}
\arguments{
\item{means}{The means of the timeseries to be generated.}

\item{stdevs}{The standard deviations of the timeseries to be generated.}

\item{correlations}{The correlation matrix of the timeseries to be generated.}

\item{n_periods}{The number of time periods for which to run the simulation.}

\item{n_simulations}{The number of simulations to perform.}
}
\value{
A list of data frames containing the simulated yields, the model
        outputs, and the DO values.
}
\description{
Generate a bond basis model. Simulate many possible outcomes for given
input parameters and calculate some derived fields based on each state.
}
