% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_basis.R
\name{build_ctd_table}
\alias{build_ctd_table}
\title{Build a table of CTDs for different contracts}
\usage{
build_ctd_table(contracts, verbose = FALSE)
}
\arguments{
\item{contracts}{A list of Bloomberg identifiers for the contracts to include}

\item{verbose}{Return additional columns}
}
\value{
Data frame of results
}
\description{
Build a table of CTDs for different contracts
}
\examples{
library(FIRVr)
build_ctd_table(c("DU", "RX"), verbose = FALSE)
}
