% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_basis.R
\name{calculate_accrued_interest}
\alias{calculate_accrued_interest}
\title{Vectorise the bond.TCF functions from jrvFinance::bond.TCF}
\usage{
calculate_accrued_interest(
  settle,
  mature,
  coupon,
  freq,
  convention,
  redemption_value = 100
)
}
\arguments{
\item{settle}{Settlement date}

\item{mature}{Maturity date}

\item{coupon}{Bond coupon as a decimal (0.1 for 10 percent)}

\item{freq}{Coupon frequency}

\item{convention}{Cashflow convention e.g. "30/360", "ACT/ACT", "ACT/360", "30/360E"}

\item{redemption_value}{Bullet redemption payment, defaults to 100}
}
\value{
A vector of accrued interest
}
\description{
Vectorise the bond.TCF functions from jrvFinance::bond.TCF
}
\examples{
library(FIRVr)
calculate_accrued_interest("2030-02-02", "2040-02-02", 0.025, 2, "ACT/ACT", 100)
}
