% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_basis.R
\name{calculate_duration}
\alias{calculate_duration}
\title{Helper function to vectorise duration calculation (bond.durations()
does not vectorise the convention argument)}
\usage{
calculate_duration(
  settle,
  mature,
  coupon,
  freq,
  yield,
  convention,
  modified,
  comp.freq,
  redemption_value = 100
)
}
\arguments{
\item{settle}{Settlement date}

\item{mature}{Maturity date}

\item{coupon}{Bond coupon as a decimal (0.1 for 10 percent)}

\item{freq}{Coupon frequency}

\item{yield}{Yield to maturity of the bond}

\item{convention}{Cashflow convention e.g. "30/360", "ACT/ACT", "ACT/360", "30/360E"}

\item{modified}{A logical, set TRUE for modified duration}

\item{comp.freq}{The frequency of compounding of the bond yield: 1 for annual, 2 for semi}

\item{redemption_value}{Bullet redemption payment, defaults to 100}
}
\value{
Calculated durations
}
\description{
Helper function to vectorise duration calculation (bond.durations()
does not vectorise the convention argument)
}
\examples{
library(FIRVr)
calculate_duration("2030-02-02", "2040-02-02", 0.025, 2, 0.212, "ACT/ACT",
                   TRUE, 2, 100)
}
