% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{conditional_densities_ou}
\alias{conditional_densities_ou}
\title{Calculate conditional densities for an Ornstein-Uhlenbeck
mean reversion model.}
\usage{
conditional_densities_ou(mu, sigma, theta, start_value, horizons, seed = NULL)
}
\arguments{
\item{mu}{The long-run mean of the process.}

\item{sigma}{The instantaneous volatility.}

\item{theta}{The mean reversion parameter.}

\item{start_value}{Value from which we expect to revert over time.}

\item{horizons}{Vector of N_business_days for which we want densities.}

\item{seed}{Optional seed for random number generation.}
}
\value{
A tibble of Horizon, Spread and Density.
}
\description{
Calculate conditional densities for an Ornstein-Uhlenbeck
mean reversion model.
}
\examples{
library(FIRVr)
densities = conditional_densities_ou(5, 5, 0.05, 20, c(5, 21))
}
