% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_connect}
\alias{db_lm_connect}
\title{Connect to a list membership database.}
\usage{
db_lm_connect(filename)
}
\arguments{
\item{filename}{The path to the SQLite database file}
}
\value{
Database connection reference.
}
\description{
The database can be created using db_lm_create_list_membership_database()
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
