% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_create_list_membership_database}
\alias{db_lm_create_list_membership_database}
\title{Create an instance of a list membership database from scratch.}
\usage{
db_lm_create_list_membership_database(filename)
}
\arguments{
\item{filename}{The path the the file which will contain the new SQLite database.}
}
\description{
Creation can only proceed it the database (file) does not already exist.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
