% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_create_membership}
\alias{db_lm_create_membership}
\title{Create a database entry to represent a list membership.}
\usage{
db_lm_create_membership(conn, list_name, isin, start_date, end_date)
}
\arguments{
\item{conn}{A database connection.}

\item{list_name}{The name of the list to be modified.}

\item{isin}{The unique security identifier.}

\item{start_date}{The start date of the membership period.}

\item{end_date}{End date for the list membership period.}
}
\description{
Adds a link between a security and a list for a specified range of dates.
List name, ISIN and start date form a primary key so that only
one entry may exist for each combination of these fields.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
