% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_disconnect}
\alias{db_lm_disconnect}
\title{Disconnect from a list membership database.}
\usage{
db_lm_disconnect(conn)
}
\arguments{
\item{conn}{A database connection.}
}
\description{
Disconnect from a list membership database.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
