% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_get_benchmarks}
\alias{db_lm_get_benchmarks}
\title{Get 'benchmark' bonds for ticker/tenor combinations.}
\usage{
db_lm_get_benchmarks(
  conn,
  ticker_groups = list(c("BKO", "OBL", "DBR"), c("FRTR")),
  list_names = NULL,
  currency = "EUR",
  benchmark_years = c(2, 5, 10),
  min_history_days = 90,
  benchmark_window_days = 365,
  target_date = Sys.Date(),
  return_all_candidates = FALSE
)
}
\arguments{
\item{conn}{Connection to the list management database}

\item{ticker_groups}{A list of arrays of tickers for which benchmarks
should be returned}

\item{list_names}{Consider only securities in these lists}

\item{benchmark_years}{Maturities for which to get benchmarks}

\item{min_history_days}{Minimum history for benchmarks in days}

\item{benchmark_window_days}{Size of window around target maturity}

\item{target_date}{Date to which benchmark maturities are calculated}
}
\value{
Tibble of benchmark data with columns TickerGroup,
        Bmk (tenor in years), SecurityDes, ISIN, MaturityDate
}
\description{
Get 'benchmark' bonds for ticker/tenor combinations.
}
\examples{
db_lm_get_benchmarks(
  conn,
  ticker_groups = list(c("BKO", "OBL", "DBR"),
                       c("FRTR")),
  list_names = c("EUR_Govt_DE", "EUR_Govt_FR"),
  benchmark_years = c(2, 5, 10),
  min_history_days = 90,
  benchmark_window_days = 365,
  target_date = Sys.Date(),
  return_all_candidates = FALSE
)

}
