% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_get_lists}
\alias{db_lm_get_lists}
\title{Get the identifiers of all membership lists in the database.}
\usage{
db_lm_get_lists(conn)
}
\arguments{
\item{conn}{A connection to a list membership database.}
}
\value{
A data frame with columns ListName and ListDesc.
}
\description{
Get the identifiers of all membership lists in the database.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
