% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_get_membership}
\alias{db_lm_get_membership}
\title{Get a list of securities that belong to a list for a specific date.}
\usage{
db_lm_get_membership(conn, list_name, date = NULL)
}
\arguments{
\item{conn}{A database connection.}

\item{list_name}{The name of the list.}

\item{date}{The date for which membership details are requested.}
}
\value{
A data frame of results.
}
\description{
Get a list of securities that belong to a list for a specific date.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
