% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_get_securities}
\alias{db_lm_get_securities}
\title{Get all securities in the database.}
\usage{
db_lm_get_securities(conn)
}
\arguments{
\item{conn}{A connection to a list membership database.}
}
\value{
A data frame with columns ISIN, SecurityDes and MaturityDate.
}
\description{
Get all securities in the database.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
