% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_update_list}
\alias{db_lm_update_list}
\title{Modify a membership list definition}
\usage{
db_lm_update_list(conn, list_name, list_desc)
}
\arguments{
\item{conn}{A database connection.}

\item{list_name}{The unique name of the list to be modified}

\item{list_desc}{A description of the list.}
}
\description{
Modify a membership list definition
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
