% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_update_membership}
\alias{db_lm_update_membership}
\title{Update an existing list membership.}
\usage{
db_lm_update_membership(conn, list_name, isin, start_date, end_date)
}
\arguments{
\item{conn}{A database connection.}

\item{list_name}{The name of the list to be modified.}

\item{isin}{The unique security identifier.}

\item{start_date}{The start date of the membership period.}

\item{end_date}{End date for the list membership period.}
}
\description{
Update an existing list membership.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
