% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_membership.R
\name{db_lm_update_security}
\alias{db_lm_update_security}
\title{Update an existing entry in the Security table.}
\usage{
db_lm_update_security(conn, isin, security_des, maturity_date)
}
\arguments{
\item{conn}{A database connection.}

\item{isin}{The unique identifier for the security.}

\item{security_des}{A description for the security e.g. Ticker Coupon Maturity}

\item{maturity_date}{The maturity date for the security.}
}
\description{
Update an existing entry in the Security table.
}
\examples{
See FIRVr/examples/ex_db_list_membership.R
}
