% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{event_study}
\alias{event_study}
\title{Create rebased event study data from timeseries and event dates.}
\usage{
event_study(df, event_dates)
}
\arguments{
\item{df}{Data frame of time series with columns Date, Series and Value.}

\item{event_dates}{Tibble of event dates with columns Series and EventDate.}
}
\value{
A data frame of time series, rebased around the event dates, with
        columns Date, Series, Value, EventDate, Time and RebasedValue
}
\description{
This method could be used for example to inspect the development
of spreads around historical bond auction dates.
}
\examples{
library(FIRVr)
library(readr)
event_dates <- readr::read_csv(
"Series,EventDate
A,2018-09-07
A,2018-08-24
B,2018-10-08
B,2018-09-10
B,2018-09-27
B,2018-12-10"
)
event_study(spread_data, event_dates)
}
