% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_basis.R
\name{extract_bond_basis_params}
\alias{extract_bond_basis_params}
\title{Estimate simulation model parameters from historical data. Allow for
volailties to be adjusted from those observed previously as e.g. yield
spread volatility can increase when bonds enter a CTD basket.}
\usage{
extract_bond_basis_params(df, volatility_adjustments = NULL)
}
\arguments{
\item{df}{A tibble with columns Date, Desc, Field and Value.}
}
\value{
A list of parameters to be used in a simulation of yield and
        yield spread developments (means, standard deviations and
        correlations), as well as field descriptions for each element.
}
\description{
Estimate simulation model parameters from historical data. Allow for
volailties to be adjusted from those observed previously as e.g. yield
spread volatility can increase when bonds enter a CTD basket.
}
