% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{fit_ou}
\alias{fit_ou}
\title{Fit an Ornstein-Uhlenbeck model to a timeseries.}
\usage{
fit_ou(series)
}
\arguments{
\item{series}{A vector of the timeseries to be fitted.}
}
\value{
A list of the model parameters, their errors and outputs from the optimisation.
}
\description{
A maximum likelihood estimate is made of the parameters
mu, theta and sigma of the model:
\eqn{dx_{t} = \theta(\mu - x_{t}) + \sigma dW_{t}}.
}
\details{
Optimisation is performed by optim, and the conversion status of
the process can be checked using the convergence value in the
returned list.
}
\examples{
library(FIRVr)
params = fit_ou(generate_ou()$Value)
}
\seealso{
\code{\link{generate_ou}}
}
