% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_basis.R
\name{generate_correlated_series}
\alias{generate_correlated_series}
\title{Generate correlated series from a multivariate normal distribution}
\usage{
generate_correlated_series(
  means,
  stdevs,
  cor_mat,
  field_descs,
  n_observations = 260,
  start_date = Sys.Date(),
  empirical = TRUE,
  seed = NULL
)
}
\arguments{
\item{means}{The means of the series.}

\item{stdevs}{The standard deviations of the series.}

\item{cor_mat}{Correlation matrix to impose.}

\item{n_observations}{Number of data points in the time series.}

\item{empirical}{Logical. If true, means and cor_mat specify the
empirical rather than population properties.}

\item{seed}{Optional seed for random number generator.}

\item{field_desc}{A label for the values to be generated.}

\item{column_prefix}{A prefix for column names in the resulting data_frame}
}
\value{
A tibble with columns Date, Desc, Field and Value
}
\description{
Generate correlated series from a multivariate normal distribution
}
\examples{
library(FIRVr)
generate_correlated_series(
  means = c(0.02672, 0.02754, 0.02758, 0.02801),
  stdevs = c(0.00070, 0.00070, 0.00070, 0.00070),
  cor_mat = matrix(
    c(1.00, 0.93, 0.91, 0.89,
      0.93, 1.00, 0.92, 0.90,
      0.91, 0.92, 1.00, 0.91,
      0.89, 0.90, 0.91, 1.00),
    ncol = 4
  ),
  field_descs = "Yield"
)
}
