% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{generate_ou}
\alias{generate_ou}
\title{Generate Ornstein-Uhlenbeck mean reverting processes}
\usage{
generate_ou(
  mu = 5,
  sigma = 5,
  theta = 0.05,
  start_value = 5,
  n_periods = 252,
  n_series = 1,
  seed = NULL
)
}
\arguments{
\item{mu}{The long-run mean of the process.}

\item{sigma}{The instantaneous volatility.}

\item{theta}{The mean reversion parameter.}

\item{start_value}{Initial value for the series.}

\item{n_periods}{The number of discrete periods for which to generate.}

\item{n_series}{The number of timeseries to generate.}

\item{seed}{Optional seed for random generation}
}
\value{
A tibble with columns Time int, Series and Spread.
}
\description{
Generates series based on the stochastic differential equation
\eqn{dx_{t} = \theta(\mu - x_{t}) + \sigma dW_{t}}.
\code{generate_ou()} can be used to generate one or many
timeseries.
}
\examples{
library(FIRVr)
df = generate_ou(n_series = 3)
}
