% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_basis.R
\name{get_ctd_dataset}
\alias{get_ctd_dataset}
\title{Get CTD data from Bloomberg}
\usage{
get_ctd_dataset(contracts, sequence = c("A", "B"))
}
\arguments{
\item{contracts}{A future codes for the contracts to include}
}
\value{
Data frame of results
}
\description{
Get CTD data from Bloomberg
}
\examples{
library(FIRVr)
get_ctd_dataset(c("DU", "TY"))
}
