% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_basis.R
\name{get_ois_curves}
\alias{get_ois_curves}
\title{Get OIS curves for one or many of USD, EUR, GBP, CAD, JPY}
\usage{
get_ois_curves(currencies, interp = FALSE, max_maturity = NULL)
}
\arguments{
\item{currencies}{A vector of currencies that may include USD, EUR, GBP, CAD, JPY}

\item{interp}{A logical flagging whether to interpolate curves}

\item{max_maturity}{Maximum maturity date of underlying points}
}
\value{
Data frame of results
}
\description{
Get OIS curves for one or many of USD, EUR, GBP, CAD, JPY
}
\examples{
library(FIRVr)
get_ois_curves(c("EUR", "USD"))
}
