% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{get_source_path}
\alias{get_source_path}
\title{Get the path for the current source file or for the project root}
\usage{
get_source_path(get_project_root = FALSE, env = parent.frame())
}
\arguments{
\item{get_project_root}{Climb backwards from the path an look for a .Rproj file}

\item{env}{The enviroment in which to run - defaults to the calling frame}
}
\value{
A file path as a string
}
\description{
Get the path for the current source file or for the project root
}
\examples{
get_source_path(get_project_root = FALSE)
}
