% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{half_life_ou}
\alias{half_life_ou}
\title{Calculate the half-life of an Ornstein-Uhlenbeck process}
\usage{
half_life_ou(theta)
}
\arguments{
\item{theta}{The estimated theta coefficient for the process}
}
\value{
The half-life.
}
\description{
Calculate the half-life of an Ornstein-Uhlenbeck process
}
