% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{ou_lik}
\alias{ou_lik}
\title{Generate a likelihood function to be used with e.g. mle() when
fitting an Ornstein-Uhlenbeck model to a timeseries.}
\usage{
ou_lik(x)
}
\arguments{
\item{x}{A timeseries of values for which O-U parameters are to be estimated.}
}
\description{
Generate a likelihood function to be used with e.g. mle() when
fitting an Ornstein-Uhlenbeck model to a timeseries.
}
