% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{plot_event_study}
\alias{plot_event_study}
\title{Plot an event study data frame that was generated with FIRVr::event_study()}
\usage{
plot_event_study(
  df,
  series = df$Series[1],
  lookback = 10,
  lookforward = 10,
  show_mean = FALSE,
  show_median = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{df}{A tibble of time series, rebased around the event dates, with
columns Date, Series, Value, EventDate, Time and RebasedValue.}

\item{series}{Which series to use for the plot.}

\item{lookback}{Number of time intervals to look back from the event.}

\item{lookforward}{Number of time intervals to look forward from the event.}

\item{show_mean}{Show the mean of the series for all events}

\item{show_median}{Show the median of the series for all events}
}
\description{
Plot an event study data frame that was generated with FIRVr::event_study()
}
\examples{
library(FIRVr)
event_dates <- readr::read_csv(
  "Series,EventDate
  A,2018-09-07
  A,2018-08-24
  B,2018-10-08
  B,2018-09-10
  B,2018-09-27
  B,2018-12-10"
)
plot_event_study(event_study(spread_data, event_dates), series = "A")
}
\seealso{
\code{\link{event_study}}
}
