% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_residuals.R
\name{plot_pca_residuals}
\alias{plot_pca_residuals}
\title{Generate a plot of pca residuals}
\usage{
plot_pca_residuals(
  df,
  n_points = 126,
  y_label = "Residual",
  show_z_score = FALSE,
  source_arg = "pca_residuals_plot",
  height = 400
)
}
\arguments{
\item{df}{A data frame of time series, with columns Date, Series, Value. A Labels column
is optional}

\item{n_points}{The number of historic points to include.}

\item{y_label}{A label for the field to be plotted.}

\item{show_z_score}{Centre and scale before performing PCA.}

\item{source_arg}{The source label to be passed to the plot_ly object.}

\item{roll_points}{The number of historich point to include in the rolling window.}
}
\description{
Generate a plot of pca residuals
}
\examples{
library(FIRVr)
plot_pca_residuals(spread_data, n_points = 63, y_label = "Residual")
}
