% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_history.R
\name{plot_spread_history}
\alias{plot_spread_history}
\title{Generate a spread history box plot}
\usage{
plot_spread_history(
  df,
  n_points = 63,
  y_label = "Spread",
  source_arg = "plot_spread_history_plot",
  height = 400
)
}
\arguments{
\item{df}{A data frame of time series, with columns Date, Series, Value. A Labels column
is optional}

\item{n_points}{Max number of historic points to include.}

\item{y_label}{A label for the field to be plotted.}

\item{source_arg}{The source label to be passed to the plot_ly object.}
}
\description{
Generate a spread history box plot
}
\examples{
library(FIRVr)
plot_spread_history(spread_data, n_points = 63, y_label = "Spread")
}
