% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_curve_fit.R
\name{price_from_curve}
\alias{price_from_curve}
\title{Calculate modelled prices for bond cashflows and a forward curve model.}
\usage{
price_from_curve(parameters, cashflows, zero_rate_model, current_date)
}
\arguments{
\item{parameters}{Vector of parameters for the Nelson-Siegel model
(beta, beta1, beta2, tau1).}

\item{cashflows}{Data frame of cashflows with columns Data, Amount, Desc.}

\item{zero_rate_model}{A function that can generate zero rates given
parameters and a time to maturity.}

\item{current_date}{The date for which the pricing takes place.}
}
\value{
A data frame with columns Desc and ModelDirtyPrice.
}
\description{
Calculate modelled prices for bond cashflows and a forward curve model.
}
\examples{
library(FIRVr)
cashflows <- tibble(
  Date = as.Date(c("2027-08-15", "2028-02-15", "2028-08-15",
                   "2029-02-15", "2029-02-15")),
  Amount = c(2.5, 2.5, 2.5, 2.5, 100),
  Desc = rep("Bond 1", 5)
)
price_from_curve(c(0.034, 0.834, -0.941, 0.957),
                 cashflows,
                 zero_coupon_nelson_siegel,
                 "2019-04-16")
}
