% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_basis.R
\name{price_treasury_bond}
\alias{price_treasury_bond}
\title{Price a US Treasury bond given either price or yield}
\usage{
price_treasury_bond(
  maturity_date,
  coupon,
  effective_date,
  yield = NA,
  price = NA
)
}
\arguments{
\item{maturity_date}{String or date specifying maturity date of bond.}

\item{coupon}{Coupon in decimal format.}

\item{effective_date}{String or date specifying when the bond started
accruing interest.}

\item{yield}{Yield in decimal format.}

\item{price}{Price.}
}
\value{
A (QuantLib) FixedRateBond object
}
\description{
Price a US Treasury bond given either price or yield
}
\examples{
library(FIRVr)
price_treasury_bond("2029-06-15", 0.025, "2009-06-15", yield = 0.021)
}
