% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_reversion.R
\name{sde_ou_coeffs_to_preferred_format}
\alias{sde_ou_coeffs_to_preferred_format}
\title{Convert sde O-U theta coefficients to preferred format.
That is: theta1, theta2, theta3 -> mu, theta, sigma.}
\usage{
sde_ou_coeffs_to_preferred_format(sde_theta_coef)
}
\arguments{
\item{sde_theta_coef}{A data frame of coefficients and standard errors}
}
\value{
A list of parameters and estimated errors in our preferred format.
}
\description{
Convert sde O-U theta coefficients to preferred format.
That is: theta1, theta2, theta3 -> mu, theta, sigma.
}
