% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spread_data}
\alias{spread_data}
\title{Sample spread data.}
\format{
A tibble: 435 x 3:
\describe{
  \item{Date}{The date for the spread.}
  \item{Series}{The series to which the spread belongs.}
  \item{Value}{The value of the spread.}
}
}
\source{
Random generation.
}
\usage{
spread_data
}
\description{
Dummy data for event_study() example.
}
\keyword{datasets}
