% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_supply.R
\name{supply_table}
\alias{supply_table}
\title{Create a list of supply events for a set of fixed income instruments.}
\usage{
supply_table(ISINs)
}
\arguments{
\item{ISINs}{A vector of 12 length character ISINs}
}
\value{
A data frame of supply events, with columns ISIN, issuer, open_date, maturity, notional, and est_PV01
}
\description{
This method can be used to get all of the open dates and sizes for 
of spreads around historical bond auction dates. It also estimates the total PV01 / DV01 
of the supply.
}
\examples{
library(FIRVr)
supply_table(c("DE0001104800", "FR0012938116"))
supply_table(c("AT0000A1XML2"))
}
