% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_basis.R
\name{to_level_and_spreads}
\alias{to_level_and_spreads}
\title{Convert a tibble of N yield series into the alternative representation of
one yield series and N-1 spread series. This is because it is convenient
to think of changes in spread volatilities for bonds in the deliverable
basket, especially when they are close to being CTD. See e.g. Huggins
and Schaller.}
\usage{
to_level_and_spreads(df)
}
\arguments{
\item{df}{A tibble with columns Date, Desc, Field and Value. The first
occurring series will be used as the level series.}
}
\value{
A tibble with columns Date, Desc, Field and Value. One level
        series and and N-1 spread series are included.
}
\description{
Convert a tibble of N yield series into the alternative representation of
one yield series and N-1 spread series. This is because it is convenient
to think of changes in spread volatilities for bonds in the deliverable
basket, especially when they are close to being CTD. See e.g. Huggins
and Schaller.
}
