% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_curve_fit.R
\name{zero_coupon_nelson_siegel}
\alias{zero_coupon_nelson_siegel}
\title{Generate zero coupon yields from a Nelson-Siegel forward rate model.}
\usage{
zero_coupon_nelson_siegel(t, parameters)
}
\arguments{
\item{t}{Time in years for required zero coupon rate.}

\item{parameters}{Vector of parameters for the Nelson-Siegel model
(beta, beta1, beta2, tau1)}
}
\value{
A zero coupon yield in decimal format
}
\description{
Generate zero coupon yields from a Nelson-Siegel forward rate model.
}
\examples{
library(FIRVr)
zero_coupon_nelson_siegel(1, c(0.034, 0.834, -0.941, 0.957))
}
